main advanced 2 commits since the previous merge:
  * de203a56 SNOW-3235556 Snowflake Custom Statement Attributes (#617)
  * 09f0a4b7 SNOW-3406388 useS3RegionalUrl / useRegionalUrl from stage info (#1205)

The only conflict was in odbc_tests/BehaviorDifferences.yaml: PR #617
appended BD#56 (SQL_SF_STMT_ATTR_LAST_QUERY_ID + SQL_SF_STMT_ATTR_MULTI_STATEMENT_COUNT)
at the same location where this PR rewrote BD#55's description for the 22015
spec rationale. Resolution: keep this PR's BD#55 description (the 22015
version that c600b4ed introduced) AND keep the new BD#56 entry verbatim from
main. No other content changed in the merge.

+++ odbc/src/api/statement.rs
@@ -1742,7 +1742,7 @@ pub fn set_stmt_attr(
             }
         }
         StmtAttr::SnowflakeMultiStatementCount => {
-            let val = value_ptr as i64;
+            let val = value_ptr as isize as i64;
             if val < -1 || val > i16::MAX as i64 {
                 return InvalidAttributeValueSnafu {
                     attribute,
